feat: refactoring read_volumes_mounts to module kubernetes_services
