+++ app/v1/modules/kubernetes_services/__init__.py
@@ -5,6 +5,7 @@ from .script_to_dict import ScriptToDictParserService
 from .script_read_volumes import ScriptVolumesReaderService
 from .script_read_env import ScriptEnvironReaderService
 from .script_read_env_from import ScriptEnvironReaderFromService
+from .script_read_volumes_mounts import ScriptVolumeMountsReaderService
 
 
 __all__ = [
@@ -12,4 +13,5 @@ __all__ = [
     "ScriptVolumesReaderService",
     "ScriptEnvironReaderFromService",
     "ScriptEnvironReaderService",
+    "ScriptVolumeMountsReaderService",
 ]
